requirements.txt
pandas==2.1.4
numpy==1.26.3
pyyaml==6.0.1
openai==1.10.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
scipy==1.11.4
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
